\name{OutlierMeth-package}
\alias{OutlierMeth-package}
\docType{package}
\title{
\packageTitle{OutlierMeth}
}
\description{
\packageDescription{OutlierMeth}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{OutlierMeth}
\packageIndices{OutlierMeth}
}
\author{
\packageAuthor{OutlierMeth}

Maintainer: \packageMaintainer{OutlierMeth}
}
\references{
NA
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }

